upadated

+++ 8feb(M1_Practice)/Section_A(OOPs)/Q3CabFare-Polymorphism(MethodOverRiding)/BaseClass.cs
@@ -2,7 +2,7 @@ using System;
 
 public class Cab
 {
-    public virtual double CalculateFare(int km)
+    public virtual double CalculateFare(double km)
     {
         return 0;
     }
